% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mokapot.R
\name{compete_max_score}
\alias{compete_max_score}
\title{Compete winners by max score within keys (one per run+spectrum_id)}
\usage{
compete_max_score(df, keys, score_col)
}
\description{
Compete winners by max score within keys (one per run+spectrum_id)
}
\keyword{internal}
