% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_stability.R
\name{dfdr_curve_stability}
\alias{dfdr_curve_stability}
\title{Stability curve across alpha values}
\usage{
dfdr_curve_stability(x, alphas, k_fixed = 10, k_sqrt_mult = 2)
}
\arguments{
\item{x}{An \code{dfdr_tbl}.}

\item{alphas}{Numeric vector of thresholds.}

\item{k_fixed}{Integer. Fixed +/-K sensitivity interval.}

\item{k_sqrt_mult}{Numeric. Multiplier for adaptive interval.}
}
\value{
A tibble of stability metrics for each alpha.
}
\description{
Stability curve across alpha values
}
