% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_stability.R
\name{dfdr_headline}
\alias{dfdr_headline}
\title{Headline stability diagnostics at a given alpha}
\usage{
dfdr_headline(x, alpha = 0.01, k_fixed = 10, k_sqrt_mult = 2)
}
\arguments{
\item{x}{dfdr_tbl}

\item{alpha}{Numeric FDR threshold between 0 and 1.}

\item{k_fixed}{integer for +/-K sensitivity (default 10)}

\item{k_sqrt_mult}{numeric multiplier for adaptive +/- ceil(mult*sqrt(D))}
}
\description{
Headline stability diagnostics at a given alpha
}
