% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BH_elasticity.R
\name{dfdr_plot_bh_elasticity}
\alias{dfdr_plot_bh_elasticity}
\title{Plot BH elasticity (Jaccard) vs alpha}
\usage{
dfdr_plot_bh_elasticity(
  el_tbl,
  xlab = "alpha (log10)",
  title = "BH elasticity: Jaccard(alpha, (1+eps)alpha)"
)
}
\arguments{
\item{el_tbl}{Output of \code{\link{dfdr_bh_elasticity}}.}

\item{xlab}{X-axis label.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot BH elasticity (Jaccard) vs alpha
}
