% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_p_density_by_decoy}
\alias{dfdr_plot_p_density_by_decoy}
\title{Plot (pseudo-)p-value density by decoy/target}
\usage{
dfdr_plot_p_density_by_decoy(
  x,
  p_max = 1,
  bw = "nrd0",
  adjust = 1,
  title = "(Pseudo-)p-value distribution (density): targets vs decoys",
  log10_x = FALSE,
  eps = 1e-300
)
}
\arguments{
\item{x}{A dfdr_tbl with column p and is_decoy.}

\item{p_max}{Max p-value displayed (default 1).}

\item{bw}{Bandwidth passed to geom_density (optional).}

\item{adjust}{Smoothing adjustment (default 1).}

\item{title}{Plot title.}

\item{log10_x}{If TRUE, plot -log10(p) instead of p (often more informative).}

\item{eps}{Lower bound to avoid Inf when log10_x=TRUE.}
}
\value{
ggplot
}
\description{
Plot (pseudo-)p-value density by decoy/target
}
