% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_pep_reliability}
\alias{dfdr_plot_pep_reliability}
\title{Plot PEP reliability}
\usage{
dfdr_plot_pep_reliability(bins_tbl, title = "PEP reliability")
}
\arguments{
\item{bins_tbl}{A PEP reliability bins table (e.g. the \verb{$bins} output of
\code{\link[=dfdr_pep_reliability]{dfdr_pep_reliability()}}) containing columns \code{pep_mean}, \code{decoy_rate}, and \code{n}.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot PEP reliability
}
