% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_pep_reliability_tdc}
\alias{dfdr_plot_pep_reliability_tdc}
\title{Plot target-focused PEP reliability (TDC-style)}
\usage{
dfdr_plot_pep_reliability_tdc(
  rel_tbl,
  title = "PEP reliability (targets; TDC-style D/T error proxy)"
)
}
\arguments{
\item{rel_tbl}{Output of dfdr_pep_reliability_tdc()}

\item{title}{Title of the plot}
}
\value{
ggplot
}
\description{
Plot target-focused PEP reliability (TDC-style)
}
