% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_scope_disagreement_matrix}
\alias{dfdr_plot_scope_disagreement_matrix}
\title{Plot scope disagreement as Jaccard overlap heatmap}
\usage{
dfdr_plot_scope_disagreement_matrix(xs, alpha = 0.01)
}
\arguments{
\item{xs}{Named list of dfdr_tbl objects}

\item{alpha}{Threshold for accepted targets}
}
\value{
A ggplot heatmap
}
\description{
Plot scope disagreement as Jaccard overlap heatmap
}
