% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_score_distributions}
\alias{dfdr_plot_score_distributions}
\title{Plot target vs decoy score distributions}
\usage{
dfdr_plot_score_distributions(x, title = "Target vs Decoy Score Distributions")
}
\arguments{
\item{x}{A dfdr_tbl object}

\item{title}{Plot title}
}
\value{
A ggplot object with target distribution in blue and decoy distribution in red
}
\description{
Plot target vs decoy score distributions
}
