% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_diann.R
\name{diann_detect_run_col}
\alias{diann_detect_run_col}
\title{Detect DIA-NN run column name}
\usage{
diann_detect_run_col(df, candidates = c("Run", "File.Name", "File.Name.Index"))
}
\description{
Detect DIA-NN run column name
}
\keyword{internal}
