% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_diann.R
\name{diann_require_cols}
\alias{diann_require_cols}
\title{Validate that required DIA-NN columns exist}
\usage{
diann_require_cols(df, cols)
}
\description{
Validate that required DIA-NN columns exist
}
\keyword{internal}
