% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzid.R
\name{mzid_score_direction}
\alias{mzid_score_direction}
\title{Determine score direction (internal)}
\usage{
mzid_score_direction(
  score_accession,
  score_direction = c("auto", "lower_better", "higher_better")
)
}
\arguments{
\item{score_accession}{Character. PSI-MS accession for the score term.}

\item{score_direction}{One of \code{"auto"}, \code{"lower_better"}, \code{"higher_better"}.}
}
\value{
\code{"lower_better"} or \code{"higher_better"}.
}
\description{
Determine score direction (internal)
}
\keyword{internal}
