% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mzid.R
\name{mzid_to_internal_score}
\alias{mzid_to_internal_score}
\title{Convert raw score to an internal "higher is better" score (internal)}
\usage{
mzid_to_internal_score(score_value, score_direction, score_accession)
}
\arguments{
\item{score_value}{Numeric vector. Raw score values.}

\item{score_direction}{Character. \code{"lower_better"} or \code{"higher_better"}.}

\item{score_accession}{Character. Score accession (used for special handling).}
}
\value{
Numeric vector. Larger values mean better matches.
}
\description{
Convert raw score to an internal "higher is better" score (internal)
}
\keyword{internal}
