% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_diann.R
\name{read_diann_parquet}
\alias{read_diann_parquet}
\title{Read a DIA-NN report.parquet}
\usage{
read_diann_parquet(path)
}
\arguments{
\item{path}{Path to report.parquet}
}
\value{
tibble with DIA-NN columns
}
\description{
Requires the 'arrow' package (in Suggests).
}
