% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectronaut_normalEG.R
\name{spectronaut_detect_run_col}
\alias{spectronaut_detect_run_col}
\title{Detect Spectronaut run column name}
\usage{
spectronaut_detect_run_col(
  df,
  candidates = c("R.FileName", "R.Condition", "File.Name")
)
}
\description{
Detect Spectronaut run column name
}
\keyword{internal}
