% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectronaut_normalEG.R
\name{spectronaut_parse_decoy}
\alias{spectronaut_parse_decoy}
\title{Parse Spectronaut boolean decoy column}
\usage{
spectronaut_parse_decoy(x)
}
\description{
Parse Spectronaut boolean decoy column
}
\keyword{internal}
