% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_spectronaut_normalEG.R
\name{spectronaut_require_cols}
\alias{spectronaut_require_cols}
\title{Validate that required Spectronaut columns exist}
\usage{
spectronaut_require_cols(df, cols)
}
\description{
Validate that required Spectronaut columns exist
}
\keyword{internal}
