% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_to_pvalue.R
\name{add_pseudo_pvalues_to_list}
\alias{add_pseudo_pvalues_to_list}
\title{Add pseudo p-values to dfdr_tbl objects that have scores but no p-values}
\usage{
add_pseudo_pvalues_to_list(xs, method = "decoy_ecdf")
}
\arguments{
\item{xs}{Named list of dfdr_tbl objects}

\item{method}{Method passed to score_to_pvalue (default "decoy_ecdf")}
}
\value{
Modified list with p column added where applicable
}
\description{
Add pseudo p-values to dfdr_tbl objects that have scores but no p-values
}
\keyword{internal}
