% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag_p_calibration.R
\name{dfdr_plot_p_calibration}
\alias{dfdr_plot_p_calibration}
\title{Plot p-value calibration (ECDF minus uniform)}
\usage{
dfdr_plot_p_calibration(ecdf_tbl, title = "P-value calibration: ECDF(p) - u")
}
\arguments{
\item{ecdf_tbl}{Output \code{$ecdf} from \code{\link{dfdr_p_calibration}}.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot p-value calibration (ECDF minus uniform)
}
