% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{dfdr_plot_pep_density_by_decoy}
\alias{dfdr_plot_pep_density_by_decoy}
\title{Plot PEP density by decoy/target}
\usage{
dfdr_plot_pep_density_by_decoy(
  x,
  pep_max = 0.5,
  bw = "nrd0",
  adjust = 1,
  title = "PEP distribution (density): targets vs decoys"
)
}
\arguments{
\item{x}{A dfdr_tbl with pep.}

\item{pep_max}{Max PEP displayed (default 0.5).}

\item{bw}{Bandwidth passed to stats::density() via geom_density (optional).}

\item{adjust}{Smoothing adjustment passed to geom_density (default 1).}

\item{title}{Plot title.}
}
\value{
ggplot
}
\description{
Plot PEP density by decoy/target
}
