% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi0_storey.R
\name{dfdr_plot_pi0}
\alias{dfdr_plot_pi0}
\title{Plot Storey pi0(lambda) curve}
\usage{
dfdr_plot_pi0(pi0_tbl, title = "Storey pi0(lambda) diagnostic")
}
\arguments{
\item{pi0_tbl}{Output \code{$pi0} from \code{\link{dfdr_pi0_storey}}.}

\item{title}{Plot title.}
}
\value{
A ggplot object.
}
\description{
Plot Storey pi0(lambda) curve
}
