% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_mokapot.R
\name{read_mokapot_psms}
\alias{read_mokapot_psms}
\title{Read mokapot PSM text outputs (targets + decoys) and combine}
\usage{
read_mokapot_psms(target_path, decoy_path)
}
\arguments{
\item{target_path}{path to *.mokapot.psms.txt (targets)}

\item{decoy_path}{path to *.mokapot.decoy.psms.txt (decoys)}
}
\value{
tibble combined
}
\description{
Read mokapot PSM text outputs (targets + decoys) and combine
}
